+++ server/src/test/java/io/squashql/spring/SquashQLApplicationTests.java
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Import;
 
 @SpringBootTest
 @Import(DatasetTestConfig.class)
-class SparkApplicationTests {
+class SquashQLApplicationTests {
 
   @Test
   void contextLoads() {
